fix: fixed routing, input and accordion
