+++ my-app/src/app/components/Accordion.tsx
@@ -9,16 +9,15 @@ type AccordionProps = {
     content: string[];
 };
 
-const Accordion: React.FC<AccordionProps> = ({ items }) => {
+const Accordion: React.FC<AccordionProps> = ({ items, content }) => {
     const [activeIndex, setActiveIndex] = useState<number | null>(null);
 
     const handleClick = (index: number) => {
         setActiveIndex(index === activeIndex ? null : index);
     };
-
     return (
         <div>
-            {items.map((item, index, content) => (
+            {items.map((item, index) => (
                 <div className="accordion-item" key={index}>
                     <span className="accordion-card"
                         onClick={() => handleClick(index)}
@@ -28,7 +27,7 @@ const Accordion: React.FC<AccordionProps> = ({ items }) => {
                     </span>
                     {activeIndex === index && (
                         <span className="accordion-content" style={{ display: 'block', marginTop: '5px' }}>
-                            {content[index]}.
+                            {content[index]}
                         </span>
                     )}
                 </div>
