+++ my-app/src/app/layout.tsx
@@ -12,7 +12,7 @@ export const metadata = {
 
 export default function RootLayout({
   children,
-}: {
+} : {
   children: React.ReactNode
 }) {
   return (
